+++ client/.eslintrc
@@ -23,7 +23,7 @@
     }
   },
   "rules": {
-    "max-len": [ 1, 80, 2],
+    "max-len": [ 1, 80, 2, { "ignoreUrls": true, "ignoreStrings": true, "ignoreTemplateLiterals": true }],
     "indent": ["error", 2],
     "one-var": 0,
     "one-var-declaration-per-line": 0,
