--- client/src/components/categoryList.js
@@ -1,29 +1,0 @@
-export default [
-  'Breakfast',
-  'Brunch',
-  'Lunch',
-  'Snacks',
-  'Appetisers',
-  'Dinner',
-  'Soups',
-  'Noodles',
-  'Rice',
-  'Pasta',
-  'Meat',
-  'Poultry',
-  'Seafood',
-  'Vegetarian',
-  'Sides',
-  'Sauces',
-  'Baking',
-  'Desserts',
-  'Drinks',
-  'Salads'
-];
-
-export const dropdown = {
-  first: ['Breakfast', 'Brunch', 'Lunch', 'Snacks', 'Appetisers'],
-  second: ['Dinner', 'Soups', 'Noodles', 'Rice', 'Pasta'],
-  third: ['Meat', 'Poultry', 'Seafood', 'Vegetarian', 'Sides'],
-  fourth: ['Sauces', 'Baking', 'Desserts', 'Drinks', 'Salads']
-};
